Move files to sci folder
