+++ src/forte_old_methods.cc
@@ -41,8 +41,8 @@
 #include "psi4/psi4-dec.h"
 #include "psi4/psifiles.h"
 
-#include "aci/aci.h"
-#include "asci.h"
+#include "sci/aci.h"
+#include "sci/asci.h"
 #include "base_classes/reference.h"
 #include "helpers/blockedtensorfactory.h"
 #include "casscf/casscf.h"
@@ -56,7 +56,7 @@
 #include "fci_mo.h"
 #include "finite_temperature.h"
 #include "helpers.h"
-#include "mrci.h"
+#include "sci/mrci.h"
 #include "mrdsrg-so/mrdsrg_so.h"
 #include "mrdsrg-so/so-mrdsrg.h"
 #include "mrdsrg-spin-adapted/dsrg_mrpt.h"
