+++ src/mrdsrg-spin-integrated/three_dsrg_mrpt2.cc
@@ -58,7 +58,7 @@
 #include "fci/fci_solver.h"
 #include "fci/fci_vector.h"
 #include "fci_mo.h"
-#include "aci/aci.h"
+#include "sci/aci.h"
 #include "three_dsrg_mrpt2.h"
 
 using namespace ambit;
