+++ src/options.cc
@@ -26,8 +26,8 @@
  * @END LICENSE
  */
 
-#include "aci/aci.h"
-#include "asci.h"
+#include "sci/aci.h"
+#include "sci/asci.h"
 #include "orbital-helpers/avas.h"
 #include "orbital-helpers/ci-no/ci-no.h"
 #include "orbital-helpers/ci-no/mrci-no.h"
