+++ src/sci/asci.cc
