+++ src/sci/asci.h
