+++ src/sci/mrci.cc
