+++ src/sci/mrpt2.cc
